# [AutoDiff] Speed up compile time by 3.5x by caching derivative function types.