--- a/include/swift/AST/AutoDiff.h
+++ b/include/swift/AST/AutoDiff.h
@@ -465,6 +465,15 @@ inline llvm::raw_ostream &operator<<(llvm::raw_ostream &s,
   return s;
 }
 
+struct SILAutoDiffDerivativeFunctionKey {
+  SILFunctionType *originalType;
+  IndexSubset *parameterIndices;
+  IndexSubset *resultIndices;
+  AutoDiffDerivativeFunctionKind kind;
+  CanGenericSignature derivativeFnGenSig;
+  bool isReabstractionThunk;
+};
+
 /// In conjunction with the original function declaration, identifies an
 /// autodiff derivative function.
 ///
@@ -599,7 +608,60 @@ template <> struct DenseMapInfo<SILAutoDiffIndices> {
   }
 };
 
-} // end namespace llvm
+using swift::SILAutoDiffDerivativeFunctionKey;
+using swift::SILFunctionType;
+using swift::IndexSubset;
+using swift::AutoDiffDerivativeFunctionKind;
+using swift::GenericSignature;
+using swift::CanGenericSignature;
+
+template <> struct DenseMapInfo<SILAutoDiffDerivativeFunctionKey> {
+
+  static bool isEqual(const SILAutoDiffDerivativeFunctionKey lhs,
+                      const SILAutoDiffDerivativeFunctionKey rhs) {
+    return lhs.originalType == rhs.originalType &&
+        lhs.parameterIndices == rhs.parameterIndices &&
+        lhs.resultIndices == rhs.resultIndices &&
+        lhs.kind == rhs.kind &&
+        lhs.derivativeFnGenSig == rhs.derivativeFnGenSig &&
+        lhs.isReabstractionThunk == rhs.isReabstractionThunk;
+  }
+
+  static inline SILAutoDiffDerivativeFunctionKey getEmptyKey() {
+    return {
+        DenseMapInfo<SILFunctionType *>::getEmptyKey(),
+        DenseMapInfo<IndexSubset *>::getEmptyKey(),
+        DenseMapInfo<IndexSubset *>::getEmptyKey(),
+        AutoDiffDerivativeFunctionKind::innerty(
+            DenseMapInfo<unsigned>::getEmptyKey()),
+        CanGenericSignature(DenseMapInfo<GenericSignature>::getEmptyKey()),
+        (bool)DenseMapInfo<unsigned>::getEmptyKey()};
+  }
+
+  static inline SILAutoDiffDerivativeFunctionKey getTombstoneKey() {
+    return {
+        DenseMapInfo<SILFunctionType *>::getTombstoneKey(),
+        DenseMapInfo<IndexSubset *>::getTombstoneKey(),
+        DenseMapInfo<IndexSubset *>::getTombstoneKey(),
+        AutoDiffDerivativeFunctionKind::innerty(
+            DenseMapInfo<unsigned>::getTombstoneKey()),
+        CanGenericSignature(DenseMapInfo<GenericSignature>::getTombstoneKey()),
+        (bool)DenseMapInfo<unsigned>::getTombstoneKey()};
+  }
+
+  static unsigned getHashValue(const SILAutoDiffDerivativeFunctionKey &Val) {
+    return hash_combine(
+        DenseMapInfo<SILFunctionType *>::getHashValue(Val.originalType),
+        DenseMapInfo<IndexSubset *>::getHashValue(Val.parameterIndices),
+        DenseMapInfo<IndexSubset *>::getHashValue(Val.resultIndices),
+        DenseMapInfo<unsigned>::getHashValue((unsigned)Val.kind.rawValue),
+        DenseMapInfo<GenericSignature>::getHashValue(Val.derivativeFnGenSig),
+        DenseMapInfo<unsigned>::getHashValue(
+            (unsigned)Val.isReabstractionThunk));
+  }
+};
+
+} // namespace llvm
 // SWIFT_ENABLE_TENSORFLOW END
 
 #endif // SWIFT_AST_AUTODIFF_H