@rxwei 2020-02-01
@swift-ci please test tensorflow