@compnerd 2020-02-01
I believe that this change may have regressed the Windows builds:

https://dev.azure.com/compnerd/swift-build/_build/results?buildId=25479&view=logs&j=0e514031-a9b0-5d19-4075-e5711178f10f&t=6491fbec-4801-5529-2943-0955e3c2f90f&l=3748