@compnerd 2020-02-01
Sadly not yet - looking into that though.  You can on master though.