@rxwei 2020-02-03
Nice graph.  This shows exactly when the slowdown occurred: It's when we switched to `SILDifferentiabilityWitness` for derivative registration.